Fix bug in `is_valid()` and Refactor some names
Changes to be committed:
	modified:   src/day04.rs

The checks inside the method have been reverted back to be closures.

Last commit to this file changed some of the checks inside the method
`is_valid`. They were no longer done inside closures. This change
however was incomplete and disrupted the functionallity of the method.
Trying to comply with the modification proved to much complex because of
nested if-else statements that did not exist previously inside closures.